자동수리시스템 UNIRX 조건문 수정 및 int형 consume을 unsigned int로 변경

+++ Assets/Script/Managers/IngameCityManager.cs
@@ -944,7 +944,7 @@ public class IngameCityManager : MonoBehaviour {
         }
 
         if (repairCount != 0) {
-            int consume = Mathf.RoundToInt(repairAmount / repairCount) / 10;
+            uint consume = (uint)(Mathf.RoundToInt(repairAmount / repairCount) / 10);
             Debug.Log(consume);
             playerController.Gold -= consume;
         }
